fixed dependency range not being calculated correctly

+++ _prog/JsonVerifier/Models/Dependency.cs
@@ -4,5 +4,5 @@ public class Dependency
 {
     public required string Task { get; set; }
     public required List<IntervalCost> Intervals { get; set; }
-    public required int UnmetCost { get; set; }
+    public required double UnmetCost { get; set; }
 }
